+++ plugin/dapp/autonomy/executor/ruleaction_test.go
@@ -272,7 +272,7 @@ func voteProposalRule(t *testing.T, env *ExecEnv, exec drivers.Driver, stateDB d
 	account := accCoin.LoadExecAccount(AddrA, address.ExecAddress(auty.AutonomyX))
 	require.Equal(t, int64(0), account.Frozen)
 	account = accCoin.LoadExecAccount(autonomyFundAddr, address.ExecAddress(auty.AutonomyX))
-	require.Equal(t, int64(proposalAmount), account.Balance)
+	require.Equal(t, proposalAmount, account.Balance)
 	// status
 	value, err := stateDB.Get(propRuleID(proposalID))
 	require.NoError(t, err)
