+++ plugin/dapp/ticket/executor/ticket_test.go
@@ -43,13 +43,13 @@ func TestTicketPrice(t *testing.T) {
 func TestCheckFork(t *testing.T) {
 	assert.Equal(t, int64(1), types.GetFork("ForkChainParamV2"))
 	p1 := types.GetP(0)
-	assert.Equal(t, int64(10000*types.Coin), p1.TicketPrice)
+	assert.Equal(t, 10000*types.Coin, p1.TicketPrice)
 	p1 = types.GetP(1)
-	assert.Equal(t, int64(3000*types.Coin), p1.TicketPrice)
+	assert.Equal(t, 3000*types.Coin, p1.TicketPrice)
 	p1 = types.GetP(2)
-	assert.Equal(t, int64(3000*types.Coin), p1.TicketPrice)
+	assert.Equal(t, 3000*types.Coin, p1.TicketPrice)
 	p1 = types.GetP(3)
-	assert.Equal(t, int64(3000*types.Coin), p1.TicketPrice)
+	assert.Equal(t, 3000*types.Coin, p1.TicketPrice)
 }
 
 func TestTicket(t *testing.T) {
